refactor: move transaction-status to common directory
