+++ frontend/src/components/common/index.ts
@@ -1,3 +1,4 @@
 export { default as NeonButton } from './neon-button';
 export { default as NeonCard } from './neon-card';
 export { default as NeonModal } from './neon-modal';
+export { default as TransactionStatus } from './transaction-status';
