+++ frontend/src/components/pet/interactions/rename-panel/index.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useMemo, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import TransactionStatus from '@components/ui/transaction-status';
+import TransactionStatus from '@components/common/transaction-status';
 import {
     getReadyPetsUnified,
     useActiveChain,
