+++ frontend/src/components/pet/send-pet-modal/index.tsx
@@ -6,7 +6,7 @@ import {
     usePetList,
     useTransferPet,
 } from '@shared/core';
-import TransactionStatus from '@components/ui/transaction-status';
+import TransactionStatus from '@components/common/transaction-status';
 import './index.css';
 
 interface SendPetModalProps {
